+++ ui/apps/everest/src/pages/database-form/database-form-schema.ts
@@ -6,7 +6,7 @@ import { ResourceSize } from './database-form-body/steps/resources/resources-ste
 import { DbWizardFormFields } from './database-form.types.ts';
 import { rfc_123_schema } from 'utils/common-validation.ts';
 import { Messages as ScheduleFormMessages } from 'components/schedule-form-dialog/schedule-form/schedule-form.messages.ts';
-import { dbVersionSchemaObject } from 'components/db-version/db-version-schema';
+import { dbVersionSchemaObject } from 'components/cluster-form/db-version/db-version-schema';
 
 const resourceToNumber = (minimum = 0) =>
   z.union([z.string().nonempty(), z.number()]).pipe(
